Fix problem with `cloudpickle`
When `cloudpickle` pickles a class that contains a `classproperty`
member, it tries to assign the object `classproperty` back to the class
raising an `AttributeError`. To avoid the problem, when a
`classproperty` is instantiated, a unique id is created. When assigning
the `classproperty` back to the class, the id is checked: if they match,
no error is raised. This shouldn't create any security issue since the
class method is not modified. This should avoid the risk of creating a
`classproperty` with the same id but different `fget`, `fset`, ...